Finished activityCreateFood Renamed layouts and acitivties with their respective new versions Fixed SQLite Connector method "createFood" -> now returns ID of created food Food ID is now "long" rather than "int"

+++ IIFYM_Tracker/src/main/java/com/example/kareem/IIFYM_Tracker/Models/Food.java
@@ -5,7 +5,7 @@ package com.example.kareem.IIFYM_Tracker.Models;
  */
 public class Food {
 
-    private int id;
+    private long id;
     private String name;
     private String brand;
     private float calories;
@@ -17,7 +17,7 @@ public class Food {
 
     // Constructor 1
     // isMeal is boolean
-    public Food(int id, String name, String brand, float calories, float carbs, float protein, float fat, int portionType, boolean isMeal) {
+    public Food(long id, String name, String brand, float calories, float carbs, float protein, float fat, int portionType, boolean isMeal) {
         this.id = id;
         this.name = name;
         this.brand = brand;
@@ -30,7 +30,7 @@ public class Food {
 
     // Constructor 2
     // isMeal is int
-    public Food(int id, String name, String brand, float calories, float carbs, float protein, float fat, int portionType, int isMeal) {
+    public Food(long id, String name, String brand, float calories, float carbs, float protein, float fat, int portionType, int isMeal) {
         this.id = id;
         this.name = name;
         this.brand = brand;
@@ -56,11 +56,11 @@ public class Food {
         this.isMeal = isMeal;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
